fix: rename remote-host workspace service part file
Flutter discovers every *_test.dart as a standalone suite. The remote-host cases are a part of workspace_service_test.dart, matching the core, removal, and handoff harness files.
